Refactoring and merging store.rs (#351) (#359)
* separating Poseidon hashing into it's own module

* move pointer types, traits, methods from store.rs to ptr.rs

* move Expression/Continuation to expr.rs and cont.rs, respectively

* refactor Ptr representation

* update subcrates

* fix ptr equality

* remove no longer meaningful test in opaque_sym

* fix benches

* cargo clippy

* add comments
