+++ src/parser.rs
@@ -2,7 +2,8 @@ use peekmore::{PeekMore, PeekMoreIterator};
 
 use crate::field::LurkField;
 use crate::package::Package;
-use crate::store::{Ptr, Store};
+use crate::ptr::Ptr;
+use crate::store::Store;
 use crate::sym::Sym;
 use crate::uint::UInt;
 use thiserror;
@@ -1205,7 +1206,7 @@ mod test {
         let ptr = s.read(input).unwrap();
         let res = s.fetch(&ptr).unwrap();
         match res {
-            crate::store::Expression::Char(c) => assert_eq!(a, c),
+            crate::expr::Expression::Char(c) => assert_eq!(a, c),
             _ => panic!("not a Char"),
         };
         let printed = res.fmt_to_string(s);
